+++ core/src/components/checkbox/checkbox.tsx
@@ -347,7 +347,12 @@ export class Checkbox implements ComponentInterface {
           </div>
           <div class="native-wrapper">
             {/* Phosphor Icons define a larger viewBox */}
-            <svg class="checkbox-icon" viewBox={theme === 'ionic' ? '0 0 256 256' : '0 0 24 24'} part="container">
+            <svg
+              class="checkbox-icon"
+              viewBox={theme === 'ionic' ? '0 0 256 256' : '0 0 24 24'}
+              part="container"
+              aria-hidden="true"
+            >
               {path}
             </svg>
           </div>