+++ packages/angular/test/apps/ng20/e2e/src/lazy/angular-version.spec.ts
@@ -1,5 +1,0 @@
-it("should be on Angular 20", () => {
-  cy.visit('/lazy');
-
-  cy.get('ion-title').contains('Angular 20');
-});